refactor: move authenticateUser and saveUserAuthToken methods from UserService to JwtTokenService
