+++ src/main/java/com/kopchak/worldoftoys/service/JwtTokenService.java
@@ -1,6 +1,8 @@
 package com.kopchak.worldoftoys.service;
 
+import com.kopchak.worldoftoys.dto.token.AccessAndRefreshTokensDto;
 import com.kopchak.worldoftoys.dto.token.AuthTokenDto;
+import com.kopchak.worldoftoys.dto.user.UserAuthDto;
 import com.kopchak.worldoftoys.model.token.AuthTokenType;
 
 import java.util.Map;
@@ -17,4 +19,5 @@ public interface JwtTokenService {
 
     AuthTokenDto refreshAccessToken(AuthTokenDto refreshTokenDto);
     void revokeAllUserAuthTokens(String username);
+    AccessAndRefreshTokensDto authenticateUser(String email);
 }
