+++ src/main/java/com/kopchak/worldoftoys/service/UserService.java
@@ -1,7 +1,5 @@
 package com.kopchak.worldoftoys.service;
 
-import com.kopchak.worldoftoys.dto.token.AccessAndRefreshTokensDto;
-import com.kopchak.worldoftoys.dto.user.UserAuthDto;
 import com.kopchak.worldoftoys.dto.user.UserRegistrationDto;
 
 public interface UserService {
@@ -10,5 +8,4 @@ public interface UserService {
     boolean isUserActivated(String email);
     boolean isNewPasswordMatchOldPassword(String resetPasswordToken, String newPassword);
     boolean isPasswordsMatch(String username, String password);
-    AccessAndRefreshTokensDto authenticateUser(UserAuthDto userAuthDto);
 }
